+++ util/net/net.go
@@ -1,6 +1,7 @@
 package net
 
 import (
+	"net"
 	"os"
 
 	"github.com/netbirdio/netbird/iface/netstack"
